+++ LAB4/Shapes/Shapes/Dot.h
@@ -4,8 +4,8 @@ class CDot : public CShape
 {
 public:
 	CDot(const coordinatesType& coordinates, const std::string& strColor);
-	float GetArea() const override;
-	float GetPerimeter() const override;
+	double GetArea() const override;
+	double GetPerimeter() const override;
 	std::string ToString() const override;
 	coordinatesType GetCoordinates() const;
